* adaptation for mobile, typo fixes

See merge request practice-summer-2019/java/hospital/frontend!38

+++ src/app/features/forms/special-shift-record-form/special-shift-record-form.component.ts
@@ -83,9 +83,11 @@ export class SpecialShiftRecordFormComponent {
   }
 
   cancelRecord(id: number) {
-    this.patientSpecialShiftService.deleteOnePatientSpecialShift(this.patientId, id).subscribe();
-    this.specialShiftService.cancelRecord(id).subscribe(data => {
-      this.reloadData();
-    });
+    if (confirm('Уверены что хотите отменить запись?')) {
+      this.patientSpecialShiftService.deleteOnePatientSpecialShift(this.patientId, id).subscribe();
+      this.specialShiftService.cancelRecord(id).subscribe(data => {
+        this.reloadData();
+      });
+    }
   }
 }
